split cmoponent methods into store
