+++ frontend/src/types/User.ts
@@ -4,3 +4,9 @@ export interface User {
 	id: number;
 	relationships: any[];
 };
+
+type RelationshipRequest = {
+	source_id: number;
+	target_id: number;
+	type: string;
+  }
